yet another batch of 64bit changes

+++ src/kernel_standalone/cpp/bottom.cc
@@ -33,7 +33,7 @@ int __gxx_personality_v0;
 int _Unwind_Resume;
 
 /* support code for new and delete */
-void *operator new(unsigned int x) {
+void *operator new(unsigned long x) {
 	return(mymalloc(x));
 }
